HACK [REACTOS] Proper SIZE_T fixes

+++ dll/win32/rpcrt4/rpc_binding.c
@@ -307,7 +307,7 @@ RPC_STATUS RPCRT4_CloseBinding(RpcBinding* Binding, RpcConnection* Connection)
 
 static LPSTR RPCRT4_strconcatA(LPSTR dst, LPCSTR src)
 {
-  DWORD len = strlen(dst), slen = strlen(src);
+  SIZE_T len = strlen(dst), slen = strlen(src);
   LPSTR ndst = HeapReAlloc(GetProcessHeap(), 0, dst, (len+slen+2)*sizeof(CHAR));
   if (!ndst)
   {
